% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LostFound.R
\docType{package}
\name{LostFound}
\alias{LostFound}
\alias{LostFound-package}
\title{Lost items on NYC MTA}
\format{
A dataframe with 251 elements
\describe{
  \item{date}{Date the item was lost}
  \item{money}{Money lost (dollars)}
  \item{phone}{Phones lost}
  \item{glasses}{Reading glasses lost}
  \item{id}{IDs lost}
  \item{credit_card}{Credit cards lost}
  \item{debit_card}{Debit cards lost}
  \item{car_keys}{Car keys lost}
  \item{house_keys}{House keys lost}
}
}
\source{
\url{http://fivethirtyeight.com/datalab/mta-new-york-lost-and-found-subway-most-common/}
}
\usage{
LostFound
}
\description{
This package contains items lost on the NYC MTA
daily in 2014 and 2015

A data set containing items lost on
the NYC MTA each day
in 2014 and 2015
}
\keyword{datasets}
